+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 - Some methods not throwing `ui-device_firmware_unsupported` when the current device firmware didn't support the method.
 
 ### Changed
+
 - @trezor/blockchain-link 2.0.0 use workers as commonjs modules in nodejs and react-native env.
+- Ethereum: EthereumSignTypedData must always have at least Trezor T parameters.
 
 # 8.2.6